Add send-miband - non-interactive utility for sending messages
It does not currently sends proper messages, it sends your message in
"caller id" field of some sort of a "missed call" message

But this is as good as we can get with currently available library.
Library needs to be extended to provide better means. I guess I will
look into Gadgetbridge code for inspiration.